ai-service/requirements.txt
flask==3.0.0
flask-cors==4.0.0
spacy==3.7.2
scikit-learn==1.4.0
PyPDF2==3.0.1
python-docx==1.1.0
numpy==1.26.3
gunicorn==21.2.0
